desk/requirements.txt
aiohappyeyeballs>=2.6.1
aiohttp>=3.13.2
aiosignal>=1.4.0
annotated-types>=0.7.0
anyio>=4.12.0
attrs>=25.4.0
beautifulsoup4>=4.14.3
blinker>=1.9.0
CacheControl>=0.14.4
cachetools>=6.2.2
certifi>=2025.11.12
cffi>=2.0.0
charset-normalizer>=3.4.4
click>=8.3.1
colorama>=0.4.6
cryptography>=46.0.3
cssselect>=1.3.0
distro>=1.9.0
feedfinder2>=0.0.4
feedparser>=6.0.12
filelock>=3.20.0
firebase_admin>=7.1.0
Flask>=3.1.2
frozenlist>=1.8.0
google-api-core>=2.28.1
google-auth>=2.45.0
google-cloud-core>=2.5.0
google-cloud-firestore>=2.21.0
google-cloud-storage>=3.6.0
google-crc32c>=1.7.1
google-genai>=1.56.0
google-resumable-media>=2.8.0
googleapis-common-protos>=1.72.0
greenlet>=3.3.0
grpcio>=1.76.0
grpcio-status>=1.76.0
h11>=0.16.0
h2>=4.3.0
hpack>=4.1.0
httpcore>=1.0.9
httpx>=0.28.1
hyperframe>=6.1.0
idna>=3.11
itsdangerous>=2.2.0
jieba3k>=0.35.1
Jinja2>=3.1.6
joblib>=1.5.2
lxml>=6.0.2
lxml_html_clean>=0.4.3
MarkupSafe>=3.0.3
msgpack>=1.1.2
multidict>=6.7.0
newspaper3k>=0.2.8
nltk>=3.9.2
pillow>=12.0.0
playwright>=1.57.0
propcache>=0.4.1
proto-plus>=1.26.1
protobuf>=6.33.2
pyasn1>=0.6.1
pyasn1_modules>=0.4.2
pycparser>=2.23
pydantic>=2.12.5
pydantic_core>=2.41.5
pyee>=13.0.0
PyJWT>=2.10.1
python-dateutil>=2.9.0.post0
python-dotenv>=1.2.1
PyYAML>=6.0.3
regex>=2025.11.3
requests>=2.32.5
requests-file>=3.0.1
rsa>=4.9.1
sgmllib3k>=1.0.0
six>=1.17.0
sniffio>=1.3.1
soupsieve>=2.8
tenacity>=9.1.2
tinysegmenter>=0.3
tldextract>=5.3.0
tqdm>=4.67.1
typing-inspection>=0.4.2
typing_extensions>=4.15.0
urllib3>=2.6.0
websockets>=15.0.1
Werkzeug>=3.1.4
yarl>=1.22.0
